test(mongodb+srv): spec and unit tests live together
